requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
phonenumbers==8.13.55
pillow==11.1.0
propcache==0.3.0
psycopg==3.2.5
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.4.6
urllib3==2.3.0
yarl==1.18.3
